+++ aofl-js-packages/__config/webpack.dev.testing.js
@@ -2,7 +2,6 @@ const path = require('path');
 const merge = require('webpack-merge');
 const common = require('./webpack.common');
 const UnitTesting = require('@aofl/unit-testing-plugin');
-const UglifyJsPlugin = require('uglifyjs-webpack-plugin');
 const webpack = require('webpack');
 
 const config = merge(common('development'), {
@@ -13,19 +12,6 @@ const config = merge(common('development'), {
   plugins: [
     new webpack.optimize.LimitChunkCountPlugin({
       maxChunks: 1
-    }),
-    new UglifyJsPlugin({
-      sourceMap: false,
-      cache: true,
-      parallel: true,
-      extractComments: true,
-      uglifyOptions: {
-        ecma: 8,
-        compress: {
-          warnings: false
-        }
-      },
-      exclude: /(custom-elements-es5-adapter)/
     }),
     new UnitTesting({
       clean: false,