+++ aofl-js-packages/component-utils/test/is-in-viewport-mixin.spec.js
@@ -1,5 +1,5 @@
 /* eslint no-invalid-this: "off" */
-import isInViewportMixin from '../src/is-in-viewport-mixin';
+import {isInViewportMixin} from '../modules/is-in-viewport-mixin';
 import {AoflElement} from '@aofl/web-components/aofl-element';
 import {render, html} from 'lit-html';